pyproject.toml
[project]
name = "assignment3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.53.1",
    "streamlit-drawable-canvas>=0.9.3",
    "pillow>=11.0.0",
    "requests>=2.32.0",
    "numpy>=2.1.0",
    "pandas>=2.2.0",
    "google-auth>=2.37.0",
]

[dependency-groups]
backend = [
    "fastapi==0.115.11",
    "uvicorn==0.34.2",
    "pydantic==2.10.6",
    "pydantic-settings==2.8.1",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "pandas==2.2.3",
    "mlflow-skinny==3.1.0",
    "python-multipart>=0.0.20",
    "python-json-logger==3.3.0",
]
dev = [
    "ipykernel>=6.29.5",
    "jupyterlab>=4.4.0",
    "jupytext>=1.17.0",
    "seaborn==0.13.2",
    "ruff==0.11.10",
    "pytest==8.3.5",
]
training = [
    "hydra-core==1.3.2",
    "hydra-optuna-sweeper==1.2.0",
    "matplotlib==3.10.1",
    "mlflow==3.1.0",
    "omegaconf>=2.3.0",
    "pandas==2.2.3",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "python-json-logger==3.3.0",
    "jsonlines==4.0.0",
    "psutil==7.0.0",
]

[tool.jupytext]
formats = "ipynb,py:percent"

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu" },
]
torchvision = [
  { index = "pytorch-cpu" },
]
